% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SCORNET-package}
\alias{SCORNET-package}
\title{SCORNET: A novel non-parametric survival curve estimator for the Electronic Health Record}
\description{
Semi-Supervised Calibration of Risk with Noisy Event Times (SCORNET) is a consistent, non-parametric
survival curve estimator that boosts efficiency over existing non-parametric estimators
by (1) utilizing unlabeled patients in a semi-supervised fashion, and (2) leveraging
information-dense engineered EHR features to maximize unlabeled set imputation precision
See Ahuja et al. (2020) BioArxiv for details
}
\keyword{package}
